#597 moved DomainContext from lookup to entity package
