+++ kernel/src/test/java/org/cristalise/kernel/test/persistency/DomainContextTest.java
@@ -24,7 +24,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import java.util.Properties;
-import org.cristalise.kernel.lookup.DomainContext;
+import org.cristalise.kernel.entity.DomainContext;
 import org.cristalise.kernel.persistency.outcome.Outcome;
 import org.cristalise.kernel.persistency.outcome.Schema;
 import org.cristalise.kernel.process.Gateway;
